generate: add --reproducible, for testing
This is for regression testing: when refactoring, there should be no
change to the generated HTML.

Eventually, we'll have a daily (or hourly) cronjob updating the git repo
and generating html for the website; timestamps will be useful for that.

It's possible that instead of having a --repoducible option, I'll add a
--add-timestamp option for that use-case and have reproducible output by
default.  For now, I just want to work on the tech debt.